move logging imports and config to log_utils, update examples with new import
