+++ distributed/tensor_parallelism/tensor_parallel_example.py
@@ -10,13 +10,8 @@ from torch.distributed.tensor.parallel import (
     RowwiseParallel,
 )
 
-import logging
-from utils import rank_log
 
-logging.basicConfig(
-    format="%(asctime)s %(message)s", datefmt="%m/%d/%Y %I:%M:%S %p", level=logging.INFO
-)
-logger = logging.getLogger(__name__)
+from log_utils import rank_log, get_logger
 
 
 """
@@ -68,6 +63,7 @@ class ToyModel(nn.Module):
 Main body of the demo of a basic version of tensor parallel by using
 PyTorch native APIs.
 """
+logger = get_logger()
 
 # create a device mesh based on the given world_size.
 _world_size = int(os.environ["WORLD_SIZE"])
